+++ operator-common/src/main/java/io/strimzi/operator/common/DefaultAdminClientProvider.java
@@ -4,21 +4,21 @@
  */
 package io.strimzi.operator.common;
 
-import io.fabric8.kubernetes.api.model.Secret;
+import io.strimzi.operator.common.model.PemAuthIdentity;
+import io.strimzi.operator.common.model.PemTrustSet;
 import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.clients.admin.AdminClientConfig;
 import org.apache.kafka.common.config.SslConfigs;
 
-import java.nio.charset.StandardCharsets;
 import java.util.Properties;
 
 /**
- * Provides the default KAfka Admin client
+ * Provides the default Kafka Admin client
  */
 public class DefaultAdminClientProvider implements AdminClientProvider {
     @Override
-    public Admin createAdminClient(String bootstrapHostnames, Secret clusterCaCertSecret, Secret keyCertSecret, String keyCertName) {
-        return createAdminClient(bootstrapHostnames, clusterCaCertSecret, keyCertSecret, keyCertName, new Properties());
+    public Admin createAdminClient(String bootstrapHostnames, PemTrustSet pemTrustSet, PemAuthIdentity pemAuthIdentity) {
+        return createAdminClient(bootstrapHostnames, pemTrustSet, pemAuthIdentity, new Properties());
     }
 
     /**
@@ -26,53 +26,38 @@ public class DefaultAdminClientProvider implements AdminClientProvider {
      *
      * 1. No TLS connection, no TLS client authentication:
      *
-     * If {@code clusterCaCertSecret}, {@code keyCertSecret} and {@code keyCertName} are null, the returned Admin Client instance
+     * If {@code pemTrustSet} and {@code pemAuthIdentity} are null, the returned Admin Client instance
      * is configured to connect to the Apache Kafka bootstrap (defined via {@code hostname}) on plain connection with no
      * TLS encryption and no TLS client authentication.
      *
      * 2. TLS connection, no TLS client authentication
      *
-     * If only {@code clusterCaCertSecret} is provided as not null, the returned Admin Client instance is configured to
+     * If only {@code pemTrustSet} is provided as not null, the returned Admin Client instance is configured to
      * connect to the Apache Kafka bootstrap (defined via {@code hostname}) on TLS encrypted connection but with no
      * TLS authentication.
      *
      * 3. TLS connection and TLS client authentication
      *
-     * If {@code clusterCaCertSecret}, {@code keyCertSecret} and {@code keyCertName} are provided as not null, the returned
+     * If {@code pemTrustSet} and {@code pemAuthIdentity} are provided as not null, the returned
      * Admin Client instance is configured to connect to the Apache Kafka bootstrap (defined via {@code hostname}) on
      * TLS encrypted connection and with TLS client authentication.
      */
     @Override
-    public Admin createAdminClient(String bootstrapHostnames, Secret clusterCaCertSecret, Secret keyCertSecret, String keyCertName, Properties config) {
-        String trustedCertificates = null;
-        String privateKey = null;
-        String certificateChain = null;
-
-        // provided Secret with cluster CA certificate for TLS encryption
-        if (clusterCaCertSecret != null) {
-            trustedCertificates = Util.certsToPemString(clusterCaCertSecret);
-        }
-
-        // provided Secret and related key for getting the private key for TLS client authentication
-        if (keyCertSecret != null && keyCertName != null && !keyCertName.isEmpty()) {
-            privateKey = new String(Util.decodeFromSecret(keyCertSecret, keyCertName + ".key"), StandardCharsets.US_ASCII);
-            certificateChain = new String(Util.decodeFromSecret(keyCertSecret, keyCertName + ".crt"), StandardCharsets.US_ASCII);
-        }
-
+    public Admin createAdminClient(String bootstrapHostnames, PemTrustSet pemTrustSet, PemAuthIdentity pemAuthIdentity, Properties config) {
         config.setProperty(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapHostnames);
 
         // configuring TLS encryption if requested
-        if (trustedCertificates != null) {
+        if (pemTrustSet != null) {
             config.setProperty(AdminClientConfig.SECURITY_PROTOCOL_CONFIG, "SSL");
             config.setProperty(SslConfigs.SSL_TRUSTSTORE_TYPE_CONFIG, "PEM");
-            config.setProperty(SslConfigs.SSL_TRUSTSTORE_CERTIFICATES_CONFIG, trustedCertificates);
+            config.setProperty(SslConfigs.SSL_TRUSTSTORE_CERTIFICATES_CONFIG, pemTrustSet.trustedCertificatesString());
         }
 
         // configuring TLS client authentication
-        if (certificateChain != null && privateKey != null) {
+        if (pemAuthIdentity != null) {
             config.setProperty(SslConfigs.SSL_KEYSTORE_TYPE_CONFIG, "PEM");
-            config.setProperty(SslConfigs.SSL_KEYSTORE_CERTIFICATE_CHAIN_CONFIG, certificateChain);
-            config.setProperty(SslConfigs.SSL_KEYSTORE_KEY_CONFIG, privateKey);
+            config.setProperty(SslConfigs.SSL_KEYSTORE_CERTIFICATE_CHAIN_CONFIG, pemAuthIdentity.pemCertificateChainString());
+            config.setProperty(SslConfigs.SSL_KEYSTORE_KEY_CONFIG, pemAuthIdentity.pemPrivateKeyString());
         }
 
         config.putIfAbsent(AdminClientConfig.METADATA_MAX_AGE_CONFIG, "30000");